Add user-carbon subgraph
This adds the `klimadao-user-carbon` subgraph, which will track balances of different carbon tokens over time.

The primary reasons this and the klima balances subgraphs are broken out from the other subgraphs is due to the fact that
all `Transfer` events must be indexed, and this can slow down the process on a larger subgraph.